requirements.txt
numpy>=1.24
pandas>=2.0
matplotlib>=3.6
seaborn>=0.12
scipy>=1.10
scikit-learn>=1.2
jupyter
pyyaml